+++ test/conftest.py
@@ -2266,6 +2266,7 @@ def picodata_expel(
     command: list[str] = [
         peer.binary_path, "expel",
         "--peer", f"pico_service@{peer.iproto_listen}",
+        "--cluster-name", target.cluster_name or "",
         "--password-file", password_file,
         "--auth-type", "chap-sha1",
         *(["--force"] if force else []),